fix: rename folder to correct name
