+++ src/main/java/com/caronte/caronte/company/CompanyRepository.java
@@ -1,4 +1,4 @@
-package com.caronte.caronte._Company;
+package com.caronte.caronte.company;
 
 import java.util.Optional;
